# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChenXuBlog",
    "version": "0.1.0"
  },
  "paths": {
...
    "/apis/v1/blog/": {
      "get": {
        "tags": ["blog"],
        "summary": "Get Blogs",
        "operationId": "get_blogs_apis_v1_blog__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 10,
              "title": "Page Size"
            },
            "description": "Page size"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in title and body",
              "title": "Search"
            },
            "description": "Search in title and body"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by user UUID",
              "title": "User Id"
            },
            "description": "Filter by user UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["blog"],
        "summary": "Create Blog",
        "operationId": "create_blog_apis_v1_blog__post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/apis/v1/blog/{id}": {
      "get": {
        "tags": ["blog"],
        "summary": "Get Blog",
        "operationId": "get_blog_apis_v1_blog__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["blog"],
        "summary": "Update Blog",
        "operationId": "update_blog_apis_v1_blog__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["blog"],
        "summary": "Delete Blog",
        "operationId": "delete_blog_apis_v1_blog__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/apis/v1/blog/{id}/like": {
      "post": {
        "tags": ["blog"],
        "summary": "Toggle Like",
        "operationId": "toggle_like_apis_v1_blog__id__like_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["blog"],
        "summary": "Get Like Status",
        "operationId": "get_like_status_apis_v1_blog__id__like_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/apis/v1/admin/blogs": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Blogs",
        "description": "获取文章列表(包含未发布的)",
        "operationId": "get_blogs_apis_v1_admin_blogs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "published",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Published"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBlogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/apis/v1/admin/blogs/{blog_id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Blog",
        "description": "删除文章",
        "operationId": "delete_blog_apis_v1_admin_blogs__blog_id__delete",
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Blog Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminBlogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminBlogResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": ["items", "total", "page", "page_size", "total_pages"],
        "title": "AdminBlogListResponse"
      },
      "AdminBlogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_uuid": {
            "type": "string",
            "title": "User Uuid"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "tags_name": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags Name"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count"
          },
          "likes_count": {
            "type": "integer",
            "title": "Likes Count"
          },
          "published": {
            "type": "boolean",
            "title": "Published"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_uuid",
          "username",
          "title",
          "cover_url",
          "tags_name",
          "created_at",
          "updated_at",
          "view_count",
          "likes_count",
          "published"
        ],
        "title": "AdminBlogResponse"
      },
...
      "BlogCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Blog title"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "description": "Blog body"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Blog tags"
          },
          "published": {
            "type": "boolean",
            "title": "Published",
            "description": "Blog published status"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url",
            "description": "Blog cover URL"
          }
        },
        "type": "object",
        "required": ["title", "body", "tags", "published"],
        "title": "BlogCreate"
      },
      "BlogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BlogResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Blog list"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "Page size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "Total pages"
          }
        },
        "type": "object",
        "required": ["items", "total", "page", "page_size", "total_pages"],
        "title": "BlogListResponse"
      },
      "BlogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Blog id"
          },
          "user_uuid": {
            "type": "string",
            "title": "User Uuid",
            "description": "User UUID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Blog title"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url",
            "description": "Blog cover URL"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "description": "Blog body"
          },
          "tags_name": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags Name",
            "description": "Blog tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Blog creation time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Blog update time"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count",
            "description": "Blog view count"
          },
          "likes_count": {
            "type": "integer",
            "title": "Likes Count",
            "description": "Blog like count"
          },
          "published": {
            "type": "boolean",
            "title": "Published",
            "description": "Blog published status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_uuid",
          "title",
          "cover_url",
          "body",
          "tags_name",
          "created_at",
          "updated_at",
          "view_count",
          "likes_count",
          "published"
        ],
        "title": "BlogResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LikeStatusResponse": {
        "properties": {
          "likes_count": {
            "type": "integer",
            "title": "Likes Count",
            "description": "Total like count"
          }
        },
        "type": "object",
        "required": ["likes_count"],
        "title": "LikeStatusResponse"
      },
      "LikeToggleResponse": {
        "properties": {
          "liked": {
            "type": "boolean",
            "title": "Liked",
            "description": "Whether the blog is liked after toggling"
          },
          "likes_count": {
            "type": "integer",
            "title": "Likes Count",
            "description": "Total like count"
          }
        },
        "type": "object",
        "required": ["liked", "likes_count"],
        "title": "LikeToggleResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}