# backend/src/backend/schema/blog.py
from datetime import datetime
from typing import List, Optional

from pydantic import BaseModel, Field


class BlogResponse(BaseModel):
    id: int = Field(..., description="Blog id")
    user_uuid: str = Field(..., description="User UUID")
    title: str = Field(..., description="Blog title")
    cover_url: Optional[str] = Field(..., description="Blog cover URL")
    body: str = Field(..., description="Blog body")
    tags_name: List[str] = Field(..., description="Blog tags")
    created_at: datetime = Field(..., description="Blog creation time")
    updated_at: datetime = Field(..., description="Blog update time")
    view_count: int = Field(..., description="Blog view count")
    likes_count: int = Field(..., description="Blog like count")
    published: bool = Field(..., description="Blog published status")

    class Config:
        from_attributes = True


class BlogCreate(BaseModel):
    title: str = Field(..., description="Blog title")
    body: str = Field(..., description="Blog body")
    tags: list[str] = Field(..., description="Blog tags")
    published: bool = Field(..., description="Blog published status")
    cover_url: Optional[str] = Field(None, description="Blog cover URL")

    class Config:
        from_attributes = True


class BlogUpdate(BaseModel):
    title: str = Field(..., description="Blog title")
    body: str = Field(..., description="Blog body")
    tags: list[str] = Field(..., description="Blog tags")
    published: bool = Field(..., description="Blog published status")
    cover_url: Optional[str] = Field(None, description="Blog cover URL")

    class Config:
        from_attributes = True


class BlogListResponse(BaseModel):
    items: List[BlogResponse] = Field(..., description="Blog list")
    total: int = Field(..., description="Total count")
    page: int = Field(..., description="Current page")
    page_size: int = Field(..., description="Page size")
    total_pages: int = Field(..., description="Total pages")

    class Config:
        from_attributes = True


class LikeToggleResponse(BaseModel):
    liked: bool = Field(..., description="Whether the blog is liked after toggling")
    likes_count: int = Field(..., description="Total like count")


class LikeStatusResponse(BaseModel):
    likes_count: int = Field(..., description="Total like count")